requirements.txt
Django==4.2
djangorestframework==3.14.0
python-decouple==3.8
django-filter==23.5
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
stripe==7.3.0
requests==2.31.0
django-jazzmin==3.0.0
